requirements.txt
msgpack-python==0.4.8
pydetector==0.3.8
